[data-table] improve virtualization performance
