+++ semcore/data-table/src/components/Body/Body.types.ts
@@ -1,6 +1,6 @@
 import { DTRow, DTRows } from './Row.types';
 import { DTColumn } from '../Head/Column.types';
-import { DTUse, VirtualScroll } from '../DataTable/DataTable.types';
+import {DataTableData, DTUse, VirtualScroll} from '../DataTable/DataTable.types';
 
 type CellRenderProps = {
   dataKey: string;
@@ -19,9 +19,8 @@ export type DataTableBodyProps = {
   renderCell?: (props: CellRenderProps) => React.ReactNode | Record<string, any>;
 };
 
-export type BodyPropsInner = {
-  rows: DTRows;
-  flatRows: DTRow[];
+export type BodyPropsInner<D extends DataTableData> = {
+  data: D;
   columns: DTColumn[];
   use: DTUse;
   compact: boolean;
@@ -36,8 +35,7 @@ export type BodyPropsInner = {
   tableContainerRef: React.RefObject<HTMLDivElement>;
   scrollTop: number;
   scrollDirection: 'down' | 'up';
-  tableRef: React.RefObject<HTMLDivElement>;
-  headerRef: React.RefObject<HTMLDivElement>;
   virtualScroll?: VirtualScroll;
   hasGroups: boolean;
+  uid: string;
 };
